Add files via upload
Hello mates! 

I am excited to share my Project - A simple dice rolling app.

The rolling dice app is a simple and intuitive tool designed for quick rolls with a clean and user-friendly interface, the app features a single button that, when clicked, rolls two virtual dice. The result of the dice roll is displayed on the screen without any animation, providing an immediate outcome. The sum of two dices is prominently shown on screen, making it easy for users to quickly see their results. Perfect for board games or just for fun, this app offers a straightforward and efficient way to roll dice with minimal fuss. This is a zip file for a rolling dice app. Extract it before use.